+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * Support filtering project and group issues by health status.
   * Support `Any`, `Today`, and `Tomorrow` issue due date filters.
   * Support sorting issue results by their title.
+  * Support filtering group issues by their type.
 
 ## Changes